+++ app/src/main/java/jp/co/soramitsu/app/root/presentation/RootViewModel.kt
@@ -230,6 +230,10 @@ class RootViewModel @Inject constructor(
     }
 
     private fun observeWalletConnectEvents() {
+        WCDelegate.registerDelegatesIfReady()
+            .onSuccess { WCDelegate.refreshConnections() }
+            .onFailure(Throwable::printStackTrace)
+
         WCDelegate.walletEvents.onEach {
             when (it) {
                 is Wallet.Model.SessionProposal -> {